Make axisartist.grid_helper_curvelinear pseudo-tests real.
The results seem to be acceptable, though the tick direction of the axis
appears to be broken.
